+++ packages/client/src/lib/socket.ts
